data_science1/requirements.txt
requests == 2.20.0
psycopg2-binary == 2.8.4
pandas == 1.2.0
nltk == 3.6.2
python-Levenshtein == 0.12.2
JPype1 == 1.2.1
scikit-learn == 0.24.1
numpy == 1.20.2
XlsxWriter == 1.4.0
strsim == 0.0.3

